+++ test/test_helper.rb
@@ -5,5 +5,8 @@ SingleCov.setup :minitest
 
 $LOAD_PATH.unshift File.expand_path("../../lib", __FILE__)
 require 'pagy'
-require_relative 'array'
+require 'rack'
+require_relative 'test_helper/array'
+require_relative 'test_helper/frontend'
+require_relative 'test_helper/backend'
 require "minitest/autorun"